Revert "handles exceptions with #to_hash properly"
This reverts commit 1b65dd77e09614d117cc41dfe25eb6192de3745f.

This commit is being reverted because the :exception key on the argument
to Notice.new is no longer an Exception decendant. As part of this
change, Notice needs to be reworked to handle that, or preferably never
instanciate a notice where the :exception key on the args is not an
Exception object.

Conflicts:

	test/notifier_test.rb