Use Capistrano's current_path.
Use Capistrano's current_path instead of release_path to allow the Cap
task to be run independent of when the release directory is created.
The release directory name is determined at Cap load time via Time.now;
however, if the release directory is not being created during this
Cap run the Airbrake Cap task tries to cd into a non-existent directory
and errors out.  Better to use current_path, which is always
correct.