Only send notices if configured with api_key
Currently if the gem is included, but not configured, it will intercept
exceptions and attempt to send them to api.airbrake.io without an
api_key.

This changes the behaviour of send_notice to only send a notice if
airbrake is configured.  This defines configured mean it has a non-blank
api_key.