Fixes airbrake/airbrake#313
Accessing the `current_user` method on ApplicationController can cause
a database access when an exception is thrown. This database access
acquires a connection from the Rails connection_pool, but there's no
clean-up to return that connection back to the pool. Eventually, the
connection pool would be empty and the application fails.