Preventing Airbrake from collecting IO objects
It appears as though IO#collect attempts to close the object. This causes
mayhem in applications that leverage Airbrake. The problem I encountered
was that Airbrake was swallowing an underlying exception thrown during a
POST request that included an attachment.

These are ugly `is_a?` checks, I'd prefer to test behavior instead of
classification. Would `respond_to?(:readlines)` be acceptable? Working
with IO, File, and Tempfile is like playing whack-a-mole.

Closes #396