Corrected ControllerMethods test that could never fail
This test was previously checking the wrong thing. It was calling a
private method to load the user (`#airbrake_current_user`), then checking
that `#id` is never called on that user. However,
`#airbrake_current_user` never calls `#id` even when a valid user is
loaded. Therefore this test could never fail.

I've "corrected" it by making it call `#airbrake_request_data` instead,
which does attempt to call `#id` (this is also the method under test for
every other test in this context).