Protected against errors while gathering user details
When airbrake reports a controller error, it attempts to get information
about `current_user`. Airbrake assumes that getting `current_user` is a
safe operation. If there is an error while loading `current_user`
airbrake does not handle the error, and crashes. This both masks the
original error with a user loading error, and prevents either error
from being reported to airbrake.

This change rescues errors that occur when calling `current_user` or
`current_member`. If this happens, the user info sent to airbrake will be
empty.

We also handle any errors raised while attempting to fetch user defined
attributes on `current_user`. In this case the erroneous attribute is
not included in the user info, but any attributes that are successfully
fetched will still be included.