Release v5.3.0
Why v5.3.0 and not v5.2.4? Well, it's because airbrake-ruby's MINOR
version was bumped. SemVer doesn't define what to do in this situation.
This gem's API stays the same. However the logic here is that users are
directly affected by this change because airbrake-ruby doesn't raise
an important exception that it used to raise.