Add try_user_signed_in method
Running current_user if a user is not signed in will attempt to authenticate the user during a request. This is problematic under the following scenario:

When a user experiences a 422 error while signing in and that error is manually reported to Airbrake, Airbrake will call current_user. By checking that the current user is signed in before calling current_user, we make sure that the user will not be authenticated preventing a user the from signing in even though their invalid authenticity tokens don't match.