circle: migrate to config version 2
I also decided to get rid of the Travis config because it never worked
how I expected. The idea is that it should run in parallel with
Circle, however in reality it would always wait for Circle to finish
its build. With Circle 2 workflows we're more flexible now, so each
job is run in a separate container making the build a little faster.

I had to comment out the rails-edge appraisal because it fails. The
fix will be addressed in future PRs.

I also locked rubocop to version 0.51. It started failing with some
unrelated errors. Rationale is the same as in:
https://github.com/airbrake/airbrake-ruby/pull/296