logger: respect level of the logger that we wrap
There was a bug where provided logger has severity ERROR but `airbrake_level`
would still be WARN (because it was hardcoded). Because of this WARN messages
would be sent to Airbrake, even though nothing would be logged.

Additionally, if you change level of the wrapped logger, airbrake_level would
get out of sync.