rack/middleware: dup middleware to avoid race conditions
Fixes #904 (can't add a new key into hash during iteration)

I believe we see this bug due to race conditions when a threaded web-server is
used. The error happens because multiple concurrent hooks write/read `@routes`.
Since we don't do anything to protect/lock this instance variable, one thread
iterates over it (possibly, on `@routes.clear`, which calls `rb_hash_foreach`),
and the other one adds some route information.

To ensure thread-safety, we dup the middleware (probably not useful now but if
we don't do it now, this will definitely bite us in the future since it's so
easy to forget). This isolates any instance variables for each request. This
strategy is [well-known][1] and [used in web frameworks such as Sinatra][2].

We also use thread-local variables via a new class called RequestStore. This is
how we share `@routes` between hooks. The old approach is, unfortunately, racey.

[1]: https://ieftimov.com/writing-rails-middleware
[2]: https://github.com/sinatra/sinatra/blob/6f15fba2790ebdf4d1215cebf425dea2ea3130ea/lib/sinatra/base.rb#L916