Deal with airbrake-ruby v3.2.6 deprecations
We are going to jump to airbrake-ruby v4 shortly after this commit because it's
not really possible to eliminate deprecation warnings and stay compatible (it's
really a lot of work that is not worth it).

In this commit we drop notifier support for our integrations. During this
process I learned that not every integration was aware of multiple notifiers and
always used the default one anyway.