rack/middleware: factor out hook installation code outside class
It was necessary to keep the hooks inside the middleware to support multiple
notifiers. Since we don't support this feautre anymore, we are feeling more
relaxed and can simplify the middleware. It was a horrible hack anyway, since we
had to have guards in `#initialize` and I'm glad it's going away.

`require` guarantees that we load the file only once, so it's safe to add
filters in the top level. We require these files from the railtie, and it also
inadvertently fixes #909 (NameError: uninitialized constant Airbrake::Rails).