rails: merge airbrake.action_controller & airbrake.http_breakdown
Initially, I wanted to make them separate, so that everything is
isolated. However, `airbrake.action_controller` loads things once
`action_controller` is ready. This is important for hooks. It's better to keep
things together, so if we need another integration, the chance of not loading
what we expect to load is lower.