Fix broken retry count logic in Sidekiq RetryableJobsFilter (#980)
It seems Sidekiq must have modified behaviour this filter relies on in recent versions. The existing code fails in two conditions:

When job['retry_count'] is nil, which happens the first time a job fails. (There is no nil check.)
When job['retry_count'] is not nil, the number represent how many retries have already been attempted. We must add 1 to this number, or retry_count will always be lower than max_attempts, even on the final retry when we want to actually notify.