generators/airbrake_initializer: log to airbrake.log
With [`performance_stats = true`][1] Airbrake generates a lot of logs now. This
causes some inconvenience in the development environment because Rails uses
`debug` severity for its logger.

We don't have a lot of places where we log but we do log frequently (on every
request we log performance data for that request). Simply removing that debug
output would solve the issue but we need it so we can debug the library.

The best decision would be using a separate file, where we can be noisy without
annoying our users. If we ever need that debug output, we can ask the users to
check `airbrake.log`.

[1]: https://github.com/airbrake/airbrake-ruby/pull/485