rails: fix broken 'airbrake:deploy'
The new suggested way to configure logger in Rails apps was added in pull
request #986. Unfortunately, there's an issue with it. We load the initialiser
on `bundle exec rake airbrake:deploy` and at this time `Rails.logger` is not
defined yet. Therefore, the deploys would fail in 100% of cases because of
NoMethodError.

To work around this issue we simply check for the existence of `Rails.logger`.
This change is simple. However, what I don't like is that we leak a lot of *our*
code to the user space. The users have to study how we setup logging. For the
sake of simplicity we want to avoid that and keep the config simple.

To achieve that, `Airbrake::Rails.logger` was added. It hides all the complexity
from the users. Probably the best part is that if users stumble upon another bug
in the logging setup, they won't need to update how the config works. Right now,
because of this bug, they have to touch their config again and this is an
oversight on my part.