circle: test Rails v4.1
It turns out some customers still use Rails v4.1. I decided to add it back to
the build matrix, so that we can be sure we still support it. We are not
building it on JRuby, 2.6 and 2.4 due to segfaults. It's not really worth
spending time on figuring out why it happens, since other Rubies pass just
fine.