rails/app: fix engine route reporting
`Rails.application.router.routes.recognize` does not return expected results
when a request is made to an engine. For example, if we have `Marketing::Engine`
and make a request to its route `/pricing`, the call would find this path:
`marketing#/`. Instead, it should be `marketing#pricing`.

In order to fix that issue we must ask every engine separately and call
`recognize` on every engine's routes. This method is supposedly slower but much
more reliable.