Test against JRuby 9.2.2.0
RSpec's output formatter breaks on jruby-9.2.0.0 because of a bug in
that jruby version (jruby/jruby#5209). This causes failures reporting
test failures in test assertions which are formatted in certain ways.
This prevents you from finding out what the test failure actually was.

To avoid that kind of problem debugging test failures, bump up the
tested JRuby version. JRuby 9.2.1.0 included a fix for that bug, but
isn't supported by circleci, so using 9.2.2.0.