rails/active_job: allow ActiveJob errors to be reported
It appears that our ActiveJob integration refuses to report errors if it's
configured in pair with Resque, DelayedJob or Sidekiq.

This code I am removing was introduced in #789. Now, 2.5 years later when I read
it, it doesn't make any sense to me. The errors are not reportred because we
explicitly guard against it.

The code is supposed to protect against the situation when we report the same
error twice (through the integration such as DelayedJob and through
ActiveJob). However after extensive testing I couldn't confirm anymore that if
this guard is removed, we report twice. I was able to confirm that we report
only once, through the ActiveJob integration.