rails/backtrace_cleaner: silence own frames
Since we rely on Rails' BacktraceCleaner to find function/file/line of the
executed SQL, sometimes the backtrace to clean can contain frames coming from
this library.

Namely, `active_record_subscriber.rb` will be returned if an SQL request is
coming from a Rails engine. In order to avoid that, we silence ourselves, so
that the cleaner can proceed to the next line and find the correct call-site
down the stack.