rails/app: ignore "catch-all" routes
Fixes #1091 (Performance - all routes appear as / with 10.0.3. Works fine in
9.5.5)

When using Rails' `recognize`, catch-all routes have higher precedence than real
routes. By skipping catch-all routes we can fall through to the next route,
which appears to be a real route.

One gotcha is that the API we use to check for "catch-all" routes is not
available on Rails 4.1 and lower (only Rails 4.2+).