rails: install all performance hooks by default
Previously, we were installing performance hooks based on the value from the
Airbrake config. If `performance_stats` is `true`, then the hooks will be
installed during initialization, if it's `false`, the hooks would never be
installed.

Now, `airbrake-ruby` supports remote configuration, which can change the value
of `performance_stats` at runtime. In order to support this feature we must
change the approach and install the hooks in either case. The
`performance_stats` check is now performed inside the hooks, since this will
allow us to enable/disable the feature during the runtime.

A pleasant effect of this change is that we no longer need to read the config
inside our railtie. This fixes #1105, which reports an issue with our existing
code where we hook into `after: :load_config_initializers`. With this new
approach we don't need to hook into that event anymore because we don't care at
what stage the `config/initializers/airbrake.rb` is loaded.