Test against Rails 6.1 and fix broken tests
Rails 6.1 deprecates rendering actions with `.` in the name, which breaks some
of our tests. We expect the `view` param in the breakdown to be present, but
it's not anymore due to this warning. Fixing the warning fixes the tests.

When I made this change, other routes started to include the `view` param as
well, therefore I added it to make other tests pass.